[level 1] Title: 평균 구하기, Time: 0.05 ms, Memory: 72.8 MB -BaekjoonHub

+++ 프로그래머스/1/12944. 평균 구하기/평균 구하기.java
@@ -1,9 +1,9 @@
 class Solution {
     public double solution(int[] arr) {
-        double sum = 0.0;
+        int sum = 0;
         for(int i : arr) {
             sum += i;
         }
-        return sum / (double) arr.length;
+        return (double) sum / arr.length;
     }
 }
